Add files via upload
 In this project I will be be creating a database of books that are stored
 using a  List. Keep track of the author, title, and publication date of each book.
Your program should have a main menu that allows the user to select from
the following: 
(1) Add a book’s author, title, and date;

 (2) Print an alphabetical list of the books sorted by author; 
and

 (3) Quit.
You must use a class to hold the data for each book. This class must hold
three string fields: one to hold the author’s name, one for the publication
date, and another to hold the book’s title. Store the entire database of
books in a list in which each list element is a book class object.
To sort the data, use the generic sort function from the <algorithm> library.
note that this requires you to define the < operator to compare two objects
of type Book so that the author field from the two books are compared.